+++ Mailer/app/Models/Generators/TldrGenerator.php
@@ -18,7 +18,7 @@ use Tomaj\NetteApi\Params\PostInputParam;
 
 class TldrGenerator implements IGenerator
 {
-    use RulesTrait;
+    use RulesTrait, TemplatesTrait;
 
     public $onSubmit;
 
@@ -94,6 +94,10 @@ class TldrGenerator implements IGenerator
         // wrap text in paragraphs
         $post = $this->helpers->wpautop($post);
         $lockedPost = $this->helpers->wpautop($lockedPost);
+
+        // parse article links
+        $post = $this->helpers->wpParseArticleLinks($post, 'https://dennikn.sk/', $this->getArticleLinkTemplateFunction());
+        $lockedPost = $this->helpers->wpParseArticleLinks($lockedPost, 'https://dennikn.sk/', $this->getArticleLinkTemplateFunction());
 
         // fix pees
         [$post, $lockedPost] = preg_replace('/<p>/is', "<p style=\"margin-bottom:26px;color:#181818;font-family:'Helvetica Neue', Helvetica, Arial, sans-serif;font-weight:normal;padding:0;text-align:left;font-size:18px;line-height:1.6;\">", [$post, $lockedPost]);